Added SX126x radio support. Added NucleoL073 and NucleoL152 board platforms support.

+++ src/radio/sx1272/sx1272.h
@@ -367,7 +367,7 @@ int16_t SX1272ReadRssi( RadioModems_t modem );
  * \param [IN]: addr Register address
  * \param [IN]: data New register value
  */
-void SX1272Write( uint8_t addr, uint8_t data );
+void SX1272Write( uint16_t addr, uint8_t data );
 
 /*!
  * \brief Reads the radio register at the specified address
@@ -375,7 +375,7 @@ void SX1272Write( uint8_t addr, uint8_t data );
  * \param [IN]: addr Register address
  * \retval data Register value
  */
-uint8_t SX1272Read( uint8_t addr );
+uint8_t SX1272Read( uint16_t addr );
 
 /*!
  * \brief Writes multiple radio registers starting at address
@@ -384,7 +384,7 @@ uint8_t SX1272Read( uint8_t addr );
  * \param [IN] buffer Buffer containing the new register's values
  * \param [IN] size   Number of registers to be written
  */
-void SX1272WriteBuffer( uint8_t addr, uint8_t *buffer, uint8_t size );
+void SX1272WriteBuffer( uint16_t addr, uint8_t *buffer, uint8_t size );
 
 /*!
  * \brief Reads multiple radio registers starting at address
@@ -393,7 +393,7 @@ void SX1272WriteBuffer( uint8_t addr, uint8_t *buffer, uint8_t size );
  * \param [OUT] buffer Buffer where to copy the registers data
  * \param [IN] size Number of registers to be read
  */
-void SX1272ReadBuffer( uint8_t addr, uint8_t *buffer, uint8_t size );
+void SX1272ReadBuffer( uint16_t addr, uint8_t *buffer, uint8_t size );
 
 /*!
  * \brief Sets the maximum payload length.
@@ -413,7 +413,7 @@ void SX1272SetMaxPayloadLength( RadioModems_t modem, uint8_t max );
 void SX1272SetPublicNetwork( bool enable );
 
 /*!
- * \brief Gets the time reuired for the board plus radio to get out of sleep.[ms]
+ * \brief Gets the time required for the board plus radio to get out of sleep.[ms]
  *
  * \retval time Radio plus board wakeup time in ms.
  */
